requirements.txt
git+https://github.com/Rapptz/discord.py
PyYAML
mysql-connector-python==8.0.12
requests